Add installation uuid
